# BookLibraryApi/bin/Release/netcoreapp3.1/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "BookLibraryApi",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Genres": {
      "get": {
        "tags": [
          "Genre"
        ],
        "operationId": "GetGenres",
        "parameters": [
          {
            "in": "query",
            "name": "PageNumber",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "PageSize",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "Fields",
            "type": "string"
          },
          {
            "in": "query",
            "name": "SearchQuery",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "head": {
        "tags": [
          "Genre"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "PageNumber",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "PageSize",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "Fields",
            "type": "string"
          },
          {
            "in": "query",
            "name": "SearchQuery",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Genre"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/GenreCreation"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "options": {
        "tags": [
          "Genre"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Genres/{id}": {
      "get": {
        "tags": [
          "Genre"
        ],
        "operationId": "GetGenre",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "query",
            "name": "fields",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "head": {
        "tags": [
          "Genre"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "query",
            "name": "fields",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Genre"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/GenreUpdate"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "options": {
        "tags": [
          "Genre"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/genres/{genreId}/Books/{bookId}/reviews": {
      "get": {
        "tags": [
          "Review"
        ],
        "operationId": "GetReviews",
        "parameters": [
          {
            "in": "path",
            "name": "genreId",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "path",
            "name": "bookId",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "query",
            "name": "OrderBy",
            "type": "string"
          },
          {
            "in": "query",
            "name": "PageNumber",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "PageSize",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "Fields",
            "type": "string"
          },
          {
            "in": "query",
            "name": "SearchQuery",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "head": {
        "tags": [
          "Review"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "genreId",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "path",
            "name": "bookId",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "query",
            "name": "OrderBy",
            "type": "string"
          },
          {
            "in": "query",
            "name": "PageNumber",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "PageSize",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "Fields",
            "type": "string"
          },
          {
            "in": "query",
            "name": "SearchQuery",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Review"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "genreId",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "path",
            "name": "bookId",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ReviewCreation"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "options": {
        "tags": [
          "Review"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "genreId",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "bookId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/genres/{genreId}/Books/{bookId}/reviews/{reviewId}": {
      "get": {
        "tags": [
          "Review"
        ],
        "operationId": "GetReview",
        "parameters": [
          {
            "in": "path",
            "name": "genreId",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "path",
            "name": "bookId",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "path",
            "name": "reviewId",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "query",
            "name": "fields",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "head": {
        "tags": [
          "Review"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "genreId",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "path",
            "name": "bookId",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "path",
            "name": "reviewId",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "query",
            "name": "fields",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Review"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "genreId",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "path",
            "name": "bookId",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "path",
            "name": "ReviewId",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ReviewUpdate"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "patch": {
        "tags": [
          "Review"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "genreId",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "path",
            "name": "bookId",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "path",
            "name": "ReviewId",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ReviewUpdateJsonPatchDocument"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "options": {
        "tags": [
          "Review"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "genreId",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "bookId",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "reviewId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
...
    "GenreCreation": {
      "required": [
        "genreName",
        "picUrl"
      ],
      "type": "object",
      "properties": {
        "description": {
          "type": "string"
        },
        "genreName": {
          "type": "string"
        },
        "picUrl": {
          "format": "uri",
          "type": "string"
        }
      }
    },
    "GenreUpdate": {
      "required": [
        "genreName",
        "picUrl"
      ],
      "type": "object",
      "properties": {
        "description": {
          "type": "string"
        },
        "genreName": {
          "type": "string"
        },
        "picUrl": {
          "format": "uri",
          "type": "string"
        }
      }
    },
    "IContractResolver": {
      "type": "object"
    },
    "OperationType": {
      "format": "int32",
      "enum": [
        0,
        1,
        2,
        3,
        4,
        5,
        6
      ],
      "type": "integer"
    },
    "ReviewCreation": {
      "required": [
        "bookRate",
        "email",
        "reviewDescription",
        "reviewerName"
      ],
      "type": "object",
      "properties": {
        "reviewerName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "reviewDescription": {
          "maxLength": 800,
          "type": "string"
        },
        "bookRate": {
          "format": "int32",
          "type": "integer"
        },
        "bookId": {
          "format": "uuid",
          "type": "string"
        }
      }
    },
    "ReviewUpdate": {
      "required": [
        "bookRate",
        "email",
        "reviewDescription",
        "reviewerName"
      ],
      "type": "object",
      "properties": {
        "reviewerName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "reviewDescription": {
          "maxLength": 800,
          "type": "string"
        },
        "bookRate": {
          "format": "int32",
          "type": "integer"
        },
        "upVote": {
          "format": "int32",
          "type": "integer"
        },
        "downVote": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "ReviewUpdateJsonPatchDocument": {
      "type": "object",
      "properties": {
        "operations": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ReviewUpdateOperation"
          },
          "readOnly": true
        },
        "contractResolver": {
          "$ref": "#/definitions/IContractResolver"
        }
      }
    },
    "ReviewUpdateOperation": {
      "type": "object",
      "properties": {
        "operationType": {
          "$ref": "#/definitions/OperationType"
        },
        "path": {
          "type": "string"
        },
        "op": {
          "type": "string"
        },
        "from": {
          "type": "string"
        },
        "value": { }
      }
    },
...
  }
}